Add files via upload
Casino Slot Machine
This is a straightforward Python slot machine game. It mimics what it's like to use a slot machine with numerous lines and symbols. To win virtual currency, match symbols across the lines in the game.

Features
Multiple Lines: The player can select from 1 to a maximum of 3 lines to bet on in this game.
Variable bet Amount: Within a predetermined range, the player can select the bet amount per line.
Random Symbol Spins: The slot machine produces spins at random with symbols that have been given various values.
Calculation of Winnings: Based on matching symbols and the bet amount, the game verifies the symbols in each line to determine the winnings.
Interactive Interface: The game has a user-friendly interface that asks for input from the player and shows the game's outcomes.
What to Do
Utilise a Python environment to run the Python script.
To begin playing, you must first make a deposit of virtual money.
To start the slot machine, press Enter.
Decide how many lines to bet on and how much to wager per line.
The symbols for each line will be seen on the slot machine.
The game will show the winnings and the winning lines if there are any winning combinations.
The balance of the player will be added to or deducted from by the net wins (winnings - total bet).
When asked to play, type 'q' to end the game.

Requirements
Python 3.x

Disclaimer
This project does not include actual money gambling; it is just intended for educational reasons. The in-game virtual currency is worthless in the real world.